MSVC projects restructuring
Until now, all the unit tests (based on googletest) were not included in
the Visual Studio project files and could not be built with VS on
Windows.

Now, the build process under VS follows the same structure as it did in
the Makefile: a library (Shitbrix.vcxproj) contains all game routines,
which are used by ShitbrixWinmain.vcxproj for the main binary and by
ShitbrixTest.vcxproj for unit testing.