Stage, MatchBuilder now use unordered associative containers.
The switch to std::unordered_{set,map} was supposed to speed up the
stuttering when blocks fall in columns with a lot of objects stacked on
top of each other.
In fact, the performance has not improved. However, the use of unordered
containers is better design because the order of the blocks in the pit
does not actually matter.