Refactor: Logic module
The Logic module is a medium-level implementation of game rules.
It contains functions to examine and modify the contents in the pit
based on a big-picture view.

The BlockDirector uses it to crank out the game state for the next tick.

This module enables more fine-grained unit testing for logic functions
that were previously internal to the BlockDirector and opaque.