Refactor Replay
With this refactoring, we move closer to getting input from a Network
component.

It reinforces the distinction between `ControllerInput`, which goes into
the `IScreen`, and `GameInput`, which is a kind of input specific to the
ingame mode. This gives us more clarity when we go to plug in more
input sources.

`GameInput` now contains the useful `game_time` field. We do not always
need it (hence the `TIME_ASAP` constant), but it is much more convenient
than always sending it with by other means.
This change is also reflected in the replay, which now primarily
produces `GameInput`s through its `ReplayEvent`s.