Copyable GameState, Input Reform
This change introduces copy operations to the new `class GameState`,
which holds multiple `Pit`s.

Game state copies are used for the `Journal` to keep checkpoints for
later reconstruction of the game state at any point in game time.

The flow of input to the game has been changed.
`GameScreen` now feeds `GameInput`s and other `ReplayEvent`s exclusively
to the `Journal`. The `Journal` plays back these events when asked.
The `GameScreen` will only display events that the `Journal` has
released.

This adaptation paves the way for the `Network` implementation, which
will be placed between `GameScreen` and `Journal`.

Replays are broken at the moment, as are certain unit tests.