Error Checking Reform
Until now, checking for errors was a barely coherent scheme of different
flavors of `SDL_assert` and a general-purpose custom `game_assert`
function.

This change abolishes the SDL assert functionality.

It adopts the standard `assert` macro for never-happens conditions and
sanity checks.

It introduces the new `enforce` function for interface contracts, which
throws an exception if the condition is not met.

It introduces several helpers to check SDL error codes, like `sdlok()`.

It introduces several different exception classes, which can carry
additional information.