Logging
Introduce the Log singleton, which is easy to use for writing formatted
messages to its output backend.

In the default `GameLoop` setup, this will be a log file.

The log integrates well with the error handling that is in place through
the new `show_error` function, which also logs all errors passed
through.