BlockDirector now operates on the whole GameState.
This change obsoletes the previous practices of keeping a reference to
the Pit, which tended to become dangling when we recalculate the state
for history inputs, and of instantiating one BlockDirector for each
Pit/player, which was redundant.

As a temporary setback, we break almost all game events. This also
affects a bunch of unit tests.