Standalone Server
There is now a new command-line option "--run", which sets the overall
run mode for the application.
This mode is "client" by default, which behaves like it did before.
Separately, "server" starts nothing but a `ServerThread` and waits for
connections.

To facilitate this, there is now a `ServerScreen` which owns the
`ServerThread`.