Network: fix ENet resource management
ENet peer and packet objects, which are allocated and given to us by the
ENet library, will no longer be arbitrarily disconnected and
deallocated.

Instead, we leave the disconnection of peers to the library when it
shuts down the host.
Also, we do not deallocate packets when the library does it upon
sending.