Network: Game Restart
As the start of the game is now controlled by the server, and can be
requested by a (not yet implemented: privileged) client, the user
regains access to the “game reset” button functionality.

Introduced an extensible network `Message` type that encapsulates
all network messages in a type + payload text format.

Introduced the new `class BasicClient`, which translates the
text-message interface to game-relevant messages and function calls.