Network: Separate BasicServer from ServerThread
By separating the high-level server functionality from the server's main
loop, the server can now adopt an architecture similar to the relation
between `GameScreen` and `BasicClient`.

This separation will also enable us to put the server on two different
clocks at the same time. This is necessary because the server must
relay messages as quickly as possible, while it keeps track of the game
at the same speed as all the clients.