GameEvents re-established
After they had been disabled during the implementation of the network
feature, game events are now implemented again.

Due to the new backtracking and synchronization feature, some new
undesirable phenomena appear. Events are duplicated, and phantom events
triggered, which are not in line with the visual game state.

In the future, there will be filters to stop some of these events from
happening.