Joystick Support
The game now accepts input from one joystick/gamepad.
The device that we get events from must be identified with the new
--joystick_number command-line option.

Other joysticks are currently ignored.
There is currently no way to remap keys.
Of the keys on the joystick, as many as possible are mapped, essentially
at random, to the game's input keys. Pressing unmapped keys might cause
the program to exit immediately.
Directional input is accepted only from the HAT input. Axis are ignored.