Network Fixes
The client now constructs the game state as soon as the server tells it
to. This prevents a crash when an input or other ingame-only message
follows immediately after.
The rest of the application is fooled by a quick&dirty state setting
that the client is merely “ready” and only starts when the screen
changes.

Also fixes a problem with the wrong exception type being thrown, for
which the error message does not appear in the log file as it should.