Graphics-free Server
Until now, even when run in server mode, the application would open a
window and display a green background.
This is not possible on server machines without any graphics devices
attached.
With this change, server mode does not try to use any of graphics,
sounds and joysticks.

The implementation now uses a global `Context` structure, which can be
configured according to the needs of the user at the time.