Events: filter duplicates
Add a new filter that can be applied to observers of game events.
This filter is used on sound and shake events. It swallows past events
that would otherwise pop up due to game state re-calculation.

Due to non-retractable game output (such as effects already seen and
sounds already heard by the user), it does not make sense to re-play
those events along the corrected history of the game record.

The current filter implementation will only allow events to pass through
in a chronologically advancing fashion.