CMakeFiles fix
Use globbing in CMakeFiles.txt to remove redundancy when the list of
source files changes. This is not recommended, but more convenient
with the parallel build systems that we use.

Link the stdc++fs library for full C++17 support. This should not be
necessary, but is currently required by gcc.

Remove the obsolete and unused winmain.cpp.