Configuration: autorecord option
The new autorecord configuration option restores the functionality to
automatically save replays at the end of a game.

This change also adds a little more error handling to the configuration
loading (although it does not help much, due to the fact that errors
will be logged only after the configuration is already loaded).