Refactor ColorSupplier
The IBlocksQueue interface and derived classes are now called
ColorSupplier.
One ColorSupplier is now responsible for generating random colors in all
contexts (new spawns and emerging from destroyed garbage).

Importantly, the GameState is no longer fixed on creating Pits with
random ColorSuppliers. This is required for testing, in which we want to
generate block colors in a deterministic way.

As an aside, the `struct GameData` from the network module is now more
widely used.