Autorecord Unfinished Games
Games were previously only saved to a replay if they finished, with one
player winning the round.

With this change, when auto-record is enabled, we also write replay of
the game when it is reset or when the game exits.

The debug functions F2 and F3 (to advance one or 8 ticks) have also been
re-established in local mode.