Fix #79: Blocks Sometimes Flicker
If a block was taken under consideration for falling down (especially
when it was resting on top of dissolving garbage), it is marked as
chaining.

Previously, this flag remained even if the block did not fall down at
all in the end. Together with the negative `m_time` of a resting block,
a graphical glitch emerged on these blocks.

With this fix, the chaining flag is cleared on blocks that end up not
falling after being marked for potential.