Fix: Autorecord on Reset
These are some changes to improve handling of replays.

If we are autorecording, we now properly do it even on reset (unlike the
previous fix, which did not actually do this).

We take care not to overwrite previous auto-replay files. If the file
already exists, we generate a name including the current second to
disambiguate. If that file too exists, we do not record - assuming that
a one-second game is worthless.

Replay-viewer mode is now nothing more than that and will immediately
exit afterwards. Reset or quit-to-menu are obsoleted. In replay mode, we
no longer create auto-replays.