Close #38: Untangle GameScreen
GameScreen no longer has a GamePhase object from a separate class
hierarchy.
Instead, the separation between intro, play and result is implemented
using an enum and a switch for a state machine. This solution is more
fitting to the modest requirements of the problem being solved.

Event relays are now owned by the Stage. This clears the GameScreen
from some unnecessary responsibilities.

For clarification of dependencies, separated stage module from state
module.