Close #69: Client: Trust my own Inputs
Inputs made for local players are now immediately processed locally on
the client, which shortcuts the round-trip to the server for better
responsiveness.

At the moment, there are no checks to ensure that only the right players
are selected. These would be only players that participate in the game
and that are controlled exclusively by this client.

In this version, there seem to be desynchronization issues with the
server. To be debugged.