Input Reform with Arbiter
In order to allow game influences based on random decisions from the
central server, the inputs are now a more diverse class of data
structures. Any server decision becomes an Input.

The random-based decision making component is called Arbiter, for which
there is a new module.

The Director and Pit no longer have the responsibility of generating the
random content.

Replays are adapted accordingly.

As the arbiter is not yet properly tested connected to the other
components, the game is currently broken.