Arbiter Unit Tests
The new test_arbiter.cpp unit contains tests for the new Arbiter
components: LocalArbiter and ServerArbiter.

Included with this change are minor fixes related to the messaging for
spawning new blocks and garbage.

Since the network module is currently not properly decoupled, the
implementation of the ServerArbiter tests is broken. The implementation
for the ServerArbiter itself is also still missing.