Game implementations exposed and testable
The derivatives of IGame - LocalGame, ServerGame and ClientGame - are
now visible in the game.hpp header to enable unit testing.

The proper solution, to expose them in a separate implementation header
and not include network.hpp via game.hpp, creates more modules, which
tastes too verbose for this small-scale project.