Draw Redesign
The Draw module and IDraw interface are now re-conceived as an
abstraction of the SDL-based implementation.

The previous concept, where a specific draw class is responsible for
drawing a specific screen, is about to overhauled. The screens will take
responsibility for drawing themselves via the new abstract interface.

This change enables us to cut out and replace the draw implementation
for testing purposes, such as the new test_screen module, which will be
functional after the completion of this redesign.

For testing the SDL draw implementation itself, this change introduces a
new demo scenario to the visualdemo application which shows off the raw
draw functionality.