Draw Redesign Adoption
The new draw structure is now in place, the old one replaced.
All Screens use it to draw themselves.

The Stage has taken over responsibility for drawing the GameScreen
contents.
For drawing the Pit(s), there is a new separate DrawPit helper class.

The first Screen test now passes as the GameScreen uses a before_reset
handler on the game object to invalidate the Stage's pointer to the game
state when it becomes dangling.