Fix Network Game Start
This change unbreaks the starting of the game from the menu screen.

Pressing the A button on the menu screen now both resets and starts the game.

The server no longer starts the game on its own. Instead, a (privileged) client must give the start signal.

Also includes small bug fixes.