Text Output Basics
The project now uses the SDL_ttf library (and underlying freetype
library) to write text on the screen. This required an SDL update.

The text drawing is handled by the new IDraw::text() function, which is
currently very simple and slow.

A simple free font from Google, Montserrat, serves as the default font
for now. The license is included in the font directory.

As a demonstration, the stage now always draws a fixed text string.