Text: Bitmap Font
This change introduces a second way of rendering text through a (very
constrained) bitmap-based font.

The advantage of the bitmap font is full control over its design, size
and extents. It also consists of two colors, which can be re-colored for
different purposes upon loading the font.

The disadvantage is that there is only one size, no kerning, and
currently no variable-width or any variability in the source bitmap at
all. It supports a subset of ASCII, with upper-case characters and most
punctuation.

The use for this text is for static strings, score etc., anything that
does not depend on user input.