Configuration: Launch Mode
The configuration now allows us to specify different setups for the
initial screen of the application.

By default, we start up into the new menu screen. For debugging
purposes, we often want to go straight into the game.

The option formerly known as `network_mode` is now `launch_mode` and
includes the new option value "menu", which is also the default.