Fix: Arbiter Conflicting Spawns
Fix a bug in the Arbiter code when spawning garbage bricks.
For constructing an input in the future, the arbiter relied on game
state information at the current moment in time. At the time when the
input takes effect, it may happen that it clashes with the changed
situation. This can lead e.g. to garbage spawning on top of each other.