Fix Arbiter Block Spawns
The Arbiter used to spawn blocks based on the current state of the pit.
For safety, the Arbiter cannot examine the current pit state for future
inputs, because the pit state may change in the mean time.

The determination which row to spawn in is now the job of the director
which raises the Starve event.