Agent: MovePossibility
The Agent module takes another step with some logic to grasp the concept
of blocks about to fall in place after some other blocks break.

By analyzing the pit with the `class MovePossibility`, the agent can see
that a certain selection of blocks are available even if they are not
yet in the same row as the target.