Agent: Match Planning
The agent now considers the available possibilities (previously
implemented) for matching any set of blocks in a 3-constellation.

It weighs the available plans using the time cost of moving the blocks
against the benefits of clearing off garbage and picks the best plan to
execute.

With this commit, the existing agent planning unit tests are now all
green. What remains is to give the agent control over a player.